requirements.txt
gradio==5.49.1
tensorflow-cpu==2.20.0
pillow==11.3.0
numpy==2.2.6
gunicorn==23.0.0
